requirements.txt
requests==2.31.0
lxml==4.9.3
schedule==1.2.0
pyyaml==6.0.1
selenium==4.15.2
webdriver-manager==4.0.1 